exploration/requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
umap-learn==0.5.6
umap-learn[plot]
spacy==3.7.5
notebook==7.2.1
ipykernel==6.29.4
transformers==4.42.1
torch==2.3.1
tensorflow==2.16.1
datasets==2.20.0
neptune==1.10.4
python-dotenv==1.0.1
kfp==2.8.0
lxml==5.2.2
beautifulsoup4==4.12.3

kfp==2.8.0
